server/requirements.txt
fastapi
uvicorn
pydantic
pynacl